+++ src/utils/logger.js
@@ -1,6 +1,7 @@
 const pino = require('pino')
 
 module.exports = pino({
+	enabled: process.env.NODE_ENV !== 'test',
 	transport: {
 		target: 'pino-pretty',
 		options: {