+++ CHANGELOG.md
@@ -1,3 +1,102 @@
+# [1.2.0-dev.14](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.13...v1.2.0-dev.14) (2024-03-09)
+
+
+### Features
+
+* 🎸 added password reset flow ([47bd7d9](https://github.com/thirst-alert/thirst-alert-be/commit/47bd7d98e2f80637b539941893bffda341f57cfa))
+
+# [1.2.0-dev.13](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.12...v1.2.0-dev.13) (2024-03-07)
+
+
+### Features
+
+* 🎸 added user settings routes ([a21ad7e](https://github.com/thirst-alert/thirst-alert-be/commit/a21ad7e664441f575cb7a21677ea9e257d161338))
+
+# [1.2.0-dev.12](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.11...v1.2.0-dev.12) (2024-3-6)
+
+
+### Features
+
+* 🎸 Need to release verifyEmail / updated verify route+test ([e213db0](https://github.com/thirst-alert/thirst-alert-be/commit/e213db0edff567a29903510cb6d04f8660a17de7))
+
+# [1.2.0-dev.11](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.10...v1.2.0-dev.11) (2024-2-26)
+
+
+### Features
+
+* 🎸 added email verification logic ([830131a](https://github.com/thirst-alert/thirst-alert-be/commit/830131a327577c57cf58e9872e6f44653199f879))
+
+# [1.2.0-dev.10](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.9...v1.2.0-dev.10) (2024-2-15)
+
+
+### Features
+
+* 🎸 temp measurement route ([9e7ff2c](https://github.com/thirst-alert/thirst-alert-be/commit/9e7ff2c786337efa780475634ab97615ef17e529))
+* 🎸 temp measurement route try two lol ([43a3f0e](https://github.com/thirst-alert/thirst-alert-be/commit/43a3f0e200b35fdc01623e86cbff67acdf2339db))
+
+# [1.2.0-dev.9](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.8...v1.2.0-dev.9) (2024-2-14)
+
+
+### Features
+
+* 🎸 final auth setup ([7ede9e2](https://github.com/thirst-alert/thirst-alert-be/commit/7ede9e22be82962b3a3c51fe9531f65f3c451b08))
+
+# [1.2.0-dev.8](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.7...v1.2.0-dev.8) (2024-2-13)
+
+
+### Features
+
+* 🎸 authentication; JWT & Passport Strategy ([7200d3d](https://github.com/thirst-alert/thirst-alert-be/commit/7200d3d5040f683abead170660f3da32b911f2fa))
+
+# [1.2.0-dev.7](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.6...v1.2.0-dev.7) (2024-2-11)
+
+
+### Features
+
+* 🎸 testing argocd-image-updater ([1589f73](https://github.com/thirst-alert/thirst-alert-be/commit/1589f73c3691df6b106642bb51fb39ba941699b9))
+
+# [1.2.0-dev.6](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.5...v1.2.0-dev.6) (2024-2-11)
+
+
+### Features
+
+* 🎸 testing argocd-image-updater ([40d5e9a](https://github.com/thirst-alert/thirst-alert-be/commit/40d5e9a05cd4d3a9ffc8e3825badc5efde65012a))
+
+# [1.2.0-dev.5](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.4...v1.2.0-dev.5) (2024-2-9)
+
+
+### Features
+
+* 🎸 temporarily disabled ssl connection for mongodb ([5d5dd8a](https://github.com/thirst-alert/thirst-alert-be/commit/5d5dd8ade95f18022df2aa16c4881e0dbc7d837e))
+
+# [1.2.0-dev.4](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.3...v1.2.0-dev.4) (2024-2-7)
+
+
+### Bug Fixes
+
+* THIS SHOULD FIX CLOUDBUILD ([98a391e](https://github.com/thirst-alert/thirst-alert-be/commit/98a391e90cc71e82e1e773708f3fb79696050ee1))
+
+# [1.2.0-dev.3](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.2...v1.2.0-dev.3) (2024-2-7)
+
+
+### Bug Fixes
+
+* maybe fix cloudbuild ([0ad4742](https://github.com/thirst-alert/thirst-alert-be/commit/0ad4742ec2bb5c8c230e84cc5339e53635aac70a))
+
+# [1.2.0-dev.2](https://github.com/thirst-alert/thirst-alert-be/compare/v1.2.0-dev.1...v1.2.0-dev.2) (2024-2-7)
+
+
+### Bug Fixes
+
+* maybe fix cloudbuild quota issue ([990afff](https://github.com/thirst-alert/thirst-alert-be/commit/990afffc2af9a8986e0f11ce2a69d30aa8a9563e))
+
+# [1.2.0-dev.1](https://github.com/thirst-alert/thirst-alert-be/compare/v1.1.3...v1.2.0-dev.1) (2024-2-3)
+
+
+### Features
+
+* 🎸 added cloudbuild ([c15a864](https://github.com/thirst-alert/thirst-alert-be/commit/c15a8646bc38c3844ed51256e575307266161133))
+
 ## [1.1.3](https://github.com/thirst-alert/thirst-alert-be/compare/v1.1.2...v1.1.3) (2024-1-29)